+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+### Changed
+- Changed `Client::get` to return `Option<Value>` in the non-error case to indicate hit vs miss.
+- Fixed a bug where reads in a particular situation would stall if the client attempted a follow-up
+  read after getting an "incomplete" protocol parse result during the last loop iteration.
+
+### Added
+- Added `Client::stats` to get a list of statistics from the server.
 
 ## [0.1.5] - 2020-09-13
 ### Changed